refactor: move manageProducts components to its own directory
