+++ app/shop/components/manageProducts/DeleteProductPrompt.tsx
