+++ app/shop/components/manageProducts/ShopProduct.tsx
